cleaned code: Data capture

+++ main/Data_Capture.h
@@ -1,19 +1,19 @@
-/*Variables that contain data from each sensor*/ 
+/*Variables that contain data from each sensor*/
 
 #ifndef DATA_CAPTURE_H
 #define DATA_CAPTURE_H
 
 
 /* Anemo meter sensor data */
-float gWindspeed = 0; 
+float gWindspeed = 0;
 
 /* BME280 */
 byte gTemperature;
-float gPressure;
-float gHumidity;
+byte gPressure;
+byte gHumidity;
 
 /* Rain Sensor */
-int gRaindata; 
+int gRaindata;
 
 /* BH1750 */
 float gLux;
@@ -21,16 +21,16 @@ float gIrradiance;
 float gPAR;
 
 /* Uv Sensor */
-float gUvindex;
+byte gUvindex;
 
 /* Wind direction */
 int gWinddirectiondata;
 
 /* Time DS3231 */
- String dateNow; 
- String timeNow;
+String dateNow;
+String timeNow;
 
- byte hourNow;
- byte minuteNow;
- 
+byte hourNow;
+byte minuteNow;
+
 #endif
